# docs/api_contract.yaml
openapi: 3.0.3
info:
  title: UEBA API Contract
  version: "1.0.0"
  description: |
    Machine-readable contract for the UEBA backend. This is a concise subset of the
    most-used endpoints: system, data ingestion, detection run, anomaly listing/resolution,
    and users/top-risk. Responses are wrapped by the project's `ok()` helper which returns
    { success, message, data, timestamp }.
servers:
  - url: /api/v1
components:
  schemas:
    ApiWrapper:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    AnomalyItem:
      type: object
      properties:
        id: { type: integer }
        uid: { type: string }
        type: { type: string }
        score: { type: number }
        risk: { type: number }
        confidence: { type: number }
        status: { type: string }
        detected_at: { type: string, format: date-time }
        evidence: { type: object, nullable: true }
    AnomaliesListData:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnomalyItem'
        count:
          type: integer
    ResolveResponseData:
      type: object
      properties:
        id: { type: integer }
        status: { type: string }
...
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
paths:
...
  /anomalies:
    get:
      summary: List anomalies (admin/analyst)
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [open, closed]
      responses:
        '200':
          description: List of anomalies (wrapped)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiWrapper'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AnomaliesListData'
  /anomalies/{anomaly_id}/resolve:
    post:
      summary: Resolve an anomaly (set status=closed)
      parameters:
        - name: anomaly_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Resolved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiWrapper'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ResolveResponseData'
...